Fixe error of int type

+++ project1-1/src/http/HttpPendingRequest.cpp
@@ -5,6 +5,6 @@ void HttpPendingRequest::pushNUmber(int number) {
     this->numbers.push_back(number);
 }
 
-void HttpPendingRequest::pushResult(int index, std::vector<int> result) {
+void HttpPendingRequest::pushResult(int index, std::vector<int64_t> result) {
     results[index] = result;
 }
